Load-import clean up temp folder
1- Fix /scratch clean up after loading active import
2- Fix /scratch clean up if some exception occurs during load-import
3- Properly display error to the user if something happens during HTTP
upload request
4- The /scratch folder can be full with non-related load-import files,
in this case, a message is displayed to the user in order to give him
some context and two solutions:
  - Delete some unused files from /scratch
  - Increase /scratch size

Test Plan:
    PASS: /scratch cleanup after importing an active load
    PASS: /scratch cleanup after importing an active load (DC proxy)
    PASS: /scratch cleanup after exception importing load
    PASS: /scratch cleanup after exception importing load (DC proxy)
    PASS: After adding more free space to /scratch, the load-import
the operation must pass without error

Related to: https://review.opendev.org/c/starlingx/distcloud/+/897563
Closes-bug: 2031941

+++ sysinv/sysinv/sysinv/sysinv/api/controllers/v1/load.py
@@ -403,6 +403,7 @@ class LoadController(rest.RestController):
 
         LOG.info("Load files: %s saved to disk." % load_files)
 
+        exception_occured = False
         try:
             new_load = pecan.request.rpcapi.start_import_load(
                 pecan.request.context,
@@ -424,10 +425,15 @@ class LoadController(rest.RestController):
                     import_type,
                 )
         except (rpc.common.Timeout, common.RemoteError) as e:
-            if os.path.isdir(constants.LOAD_FILES_STAGING_DIR):
-                shutil.rmtree(constants.LOAD_FILES_STAGING_DIR)
+            exception_occured = True
             error = e.value if hasattr(e, 'value') else str(e)
             raise wsme.exc.ClientSideError(error)
+        except Exception:
+            exception_occured = True
+            raise
+        finally:
+            if exception_occured and os.path.isdir(constants.LOAD_FILES_STAGING_DIR):
+                shutil.rmtree(constants.LOAD_FILES_STAGING_DIR)
 
         load_data = new_load.as_dict()
         LOG.info("Load import request validated, returning new load data: %s"
